version 3.1.3

+++ app/src/main/java/com/example/plantea/presentacion/actividades/EncryptionUtils.kt
@@ -26,7 +26,7 @@ class EncryptionUtils {
 
             val sb = StringBuilder()
             for (b in cipherText) {
-                sb.append(b.toChar())
+                sb.append(b.toInt().toChar())
             }
 
             return cipherText.toString(Charsets.UTF_8)
@@ -41,7 +41,7 @@ class EncryptionUtils {
 
             val sb = StringBuilder()
             for (b in cipherText) {
-                sb.append(b.toChar())
+                sb.append(b.toInt().toChar())
             }
             return cipherText.toString(Charsets.UTF_8)
